refactor: merge the Resolver between the Gateway and the Router
