+++ node/bft/src/helpers/mod.rs
@@ -37,9 +37,6 @@ pub use proposal_cache::*;
 pub mod ready;
 pub use ready::*;
 
-pub mod resolver;
-pub use resolver::*;
-
 pub mod signed_proposals;
 pub use signed_proposals::*;
