+++ node/router/src/handshake.rs
@@ -134,7 +134,7 @@ impl<N: Network> Router<N> {
         if let Some(addr) = listener_addr {
             if let Ok(ref cr) = handshake_result {
                 if let Some(peer) = self.peer_pool.write().get_mut(&addr) {
-                    self.resolver.write().insert_peer(peer.listener_addr(), peer_addr);
+                    self.resolver.write().insert_peer(peer.listener_addr(), peer_addr, cr.address);
                     peer.upgrade_to_connected(peer_addr, cr.listener_port, cr.address, cr.node_type, cr.version);
                 }
                 #[cfg(feature = "metrics")]
