+++ node/src/client/router.rs
@@ -66,7 +66,7 @@ impl<N: Network, C: ConsensusStorage<N>> Handshake for Client<N, C> {
 impl<N: Network, C: ConsensusStorage<N>> OnConnect for Client<N, C> {
     async fn on_connect(&self, peer_addr: SocketAddr) {
         // Resolve the peer address to the listener address.
-        let Some(peer_ip) = self.router.resolve_to_listener(&peer_addr) else { return };
+        let Some(peer_ip) = self.router.resolve_to_listener(peer_addr) else { return };
         // If it's a bootstrap peer, first request its peers.
         if bootstrap_peers::<N>(self.router.is_dev()).contains(&peer_ip) {
             self.router().send(peer_ip, Message::PeerRequest(PeerRequest));
@@ -80,7 +80,7 @@ impl<N: Network, C: ConsensusStorage<N>> OnConnect for Client<N, C> {
 impl<N: Network, C: ConsensusStorage<N>> Disconnect for Client<N, C> {
     /// Any extra operations to be performed during a disconnect.
     async fn handle_disconnect(&self, peer_addr: SocketAddr) {
-        if let Some(peer_ip) = self.router.resolve_to_listener(&peer_addr) {
+        if let Some(peer_ip) = self.router.resolve_to_listener(peer_addr) {
             self.sync.remove_peer(&peer_ip);
             self.router.remove_connected_peer(peer_ip);
         }
@@ -123,7 +123,7 @@ impl<N: Network, C: ConsensusStorage<N>> Client<N, C> {
         // Process the message. Disconnect if the peer violated the protocol.
         if let Err(error) = self.inbound(peer_addr, message).await {
             warn!("Failed to process inbound message from '{peer_addr}' - {error}");
-            if let Some(peer_ip) = self.router().resolve_to_listener(&peer_addr) {
+            if let Some(peer_ip) = self.router().resolve_to_listener(peer_addr) {
                 warn!("Disconnecting from '{peer_ip}' for protocol violation");
                 self.router().send(peer_ip, Message::Disconnect(DisconnectReason::ProtocolViolation.into()));
                 // Disconnect from this peer.
