+++ node/src/validator/router.rs
@@ -66,7 +66,7 @@ where
 {
     async fn on_connect(&self, peer_addr: SocketAddr) {
         // Resolve the peer address to the listener address.
-        let Some(peer_ip) = self.router.resolve_to_listener(&peer_addr) else { return };
+        let Some(peer_ip) = self.router.resolve_to_listener(peer_addr) else { return };
         // Send the first `Ping` message to the peer.
         self.ping.on_peer_connected(peer_ip);
     }
@@ -76,7 +76,7 @@ where
 impl<N: Network, C: ConsensusStorage<N>> Disconnect for Validator<N, C> {
     /// Any extra operations to be performed during a disconnect.
     async fn handle_disconnect(&self, peer_addr: SocketAddr) {
-        if let Some(peer_ip) = self.router.resolve_to_listener(&peer_addr) {
+        if let Some(peer_ip) = self.router.resolve_to_listener(peer_addr) {
             self.sync.remove_peer(&peer_ip);
             self.router.remove_connected_peer(peer_ip);
         }
@@ -119,7 +119,7 @@ impl<N: Network, C: ConsensusStorage<N>> Validator<N, C> {
         // Process the message. Disconnect if the peer violated the protocol.
         if let Err(error) = self.inbound(peer_addr, message).await {
             warn!("Failed to process inbound message from '{peer_addr}' - {error}");
-            if let Some(peer_ip) = self.router().resolve_to_listener(&peer_addr) {
+            if let Some(peer_ip) = self.router().resolve_to_listener(peer_addr) {
                 warn!("Disconnecting from '{peer_ip}' for protocol violation");
                 self.router().send(peer_ip, Message::Disconnect(DisconnectReason::ProtocolViolation.into()));
                 // Disconnect from this peer.
